allow to not manage mountpoint

+++ manifests/volumes/vol.pp
@@ -11,7 +11,8 @@ define g_server::volumes::vol (
   Optional[String] $thinpool = undef,
   Optional[String] $mountpoint_user = undef,
   Optional[String] $mountpoint_group = undef,
-  Optional[String] $mountpoint_mode = undef
+  Optional[String] $mountpoint_mode = undef,
+  Boolean $manage_mountpoint = true
 ){
   
   lvm::logical_volume { $lv_name:
@@ -32,17 +33,19 @@ define g_server::volumes::vol (
   
   if $ensure == 'present' {
     
-    file { $mountpoint:
-      ensure => $ensure?{
-        'present' => directory,
-        default => $ensure
-      },
-      backup => false,
-      force => true,
-      recurse => false,
-      owner => $mountpoint_user,
-      group => $mountpoint_group,
-      mode => $mountpoint_mode
+    if $manage_mountpoint {
+      file { $mountpoint:
+        ensure => $ensure?{
+          'present' => directory,
+          default => $ensure
+        },
+        backup => false,
+        force => true,
+        recurse => false,
+        owner => $mountpoint_user,
+        group => $mountpoint_group,
+        mode => $mountpoint_mode
+      }
     }
     
     /*File[$mountpoint]
@@ -52,15 +55,18 @@ define g_server::volumes::vol (
     Exec <| title=="ensure mountpoint '${mountpoint}' exists" |> {
       unless => "true",
     }
-    # fix for dependency cycle when ensure=>absent and file is autorequiring parent
-    # https://tickets.puppetlabs.com/browse/PUP-2451
-    exec { "g_server remove vol mountpoint ${mountpoint}":
-      path     => '/usr/bin:/usr/sbin:/bin',
-      command  => "rmdir '${mountpoint}'",
-      onlyif   => "test -d '${mountpoint}'"
-    }
     
-    Mount[$mountpoint]
-    ->Exec["g_server remove vol mountpoint ${mountpoint}"]
+    if $manage_mountpoint {
+      # fix for dependency cycle when ensure=>absent and file is autorequiring parent
+      # https://tickets.puppetlabs.com/browse/PUP-2451
+      exec { "g_server remove vol mountpoint ${mountpoint}":
+        path     => '/usr/bin:/usr/sbin:/bin',
+        command  => "rmdir '${mountpoint}'",
+        onlyif   => "test -d '${mountpoint}'"
+      }
+      
+      Mount[$mountpoint]
+      ->Exec["g_server remove vol mountpoint ${mountpoint}"]
+    }
   }
 }
